Added DrawerAdapter class
Moved drawer settings to another class in package components
