+++ app/src/main/java/com/blstream/studybox/activities/DecksActivity.java
@@ -1,18 +1,17 @@
 package com.blstream.studybox.activities;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.Toast;
 
+import com.blstream.studybox.components.DrawerAdapter;
 import com.blstream.studybox.R;
 import com.blstream.studybox.decks_view.DecksAdapter;
 import com.blstream.studybox.decks_view.DecksPresenter;
@@ -25,7 +24,7 @@ import butterknife.BindInt;
 import butterknife.ButterKnife;
 
 public class DecksActivity extends MvpActivity<DecksView, DecksPresenter>
-        implements DecksView, DecksAdapter.ClickListener, NavigationView.OnNavigationItemSelectedListener {
+        implements DecksView, DecksAdapter.ClickListener {
 
     @Bind(R.id.decks_recycler_view)
     RecyclerView recyclerView;
@@ -43,23 +42,23 @@ public class DecksActivity extends MvpActivity<DecksView, DecksPresenter>
     @Bind(R.id.drawer_layout_decks)
     DrawerLayout drawerLayout;
 
+    DrawerAdapter drawerAdapter;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_decks);
         initView();
-        setUpRecyclerView();
         loadData();
     }
+
     private void initView() {
         ButterKnife.bind(this);
         setSupportActionBar(toolbar);
-        navigationView.setNavigationItemSelectedListener(this);
-
-        ActionBarDrawerToggle drawerToggle = new ActionBarDrawerToggle(
-                this, drawerLayout, toolbar, R.string.nav_open_drawer, R.string.nav_close_drawer);
-        drawerLayout.addDrawerListener(drawerToggle);
-        drawerToggle.syncState();
+        Context context = getApplicationContext();      //only For testing
+        drawerAdapter = new DrawerAdapter(navigationView, drawerLayout, toolbar, context);
+        drawerAdapter.attachDrawer();
+        setUpRecyclerView();
     }
 
     private void setUpRecyclerView() {
@@ -93,36 +92,10 @@ public class DecksActivity extends MvpActivity<DecksView, DecksPresenter>
     }
 
     @Override
-    public boolean onNavigationItemSelected(MenuItem item) {
-        item.setChecked(true);
-
-        int id = item.getItemId();
-        drawerLayout.closeDrawer(GravityCompat.START);
-
-        switch (id) {
-            case R.id.my_account:
-                break;
-            case R.id.my_decks:
-                break;
-            case R.id.create_flashcard:
-                break;
-            case R.id.show_deck:
-                break;
-            case R.id.statistics:
-                break;
-            case R.id.logout:
-                break;
-        }
-
-        // Only for testing
-        Toast.makeText(this, "Selected: " + item.getTitle(), Toast.LENGTH_SHORT).show();
-
-        return true;
-    }
-    @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
         getMenuInflater().inflate(R.menu.search_toolbar_menu, menu);
+
         return true;
     }
 }
