+++ app/src/main/java/com/blstream/studybox/activities/ExamActivity.java
@@ -1,18 +1,16 @@
 package com.blstream.studybox.activities;
 
+import android.content.Context;
 import android.os.Bundle;
 import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
-import android.view.MenuItem;
 import android.view.View;
 import android.widget.TextView;
-import android.widget.Toast;
 
+import com.blstream.studybox.components.DrawerAdapter;
 import com.blstream.studybox.R;
 import com.blstream.studybox.exam_view.DeckPagerAdapter;
 import com.blstream.studybox.exam_view.DeckViewPager;
@@ -23,7 +21,7 @@ import java.util.List;
 import butterknife.Bind;
 import butterknife.ButterKnife;
 
-public class ExamActivity extends AppCompatActivity implements NavigationView.OnNavigationItemSelectedListener {
+public class ExamActivity extends AppCompatActivity {
 
     @Bind(R.id.deckName)
     public TextView deckName;
@@ -51,6 +49,7 @@ public class ExamActivity extends AppCompatActivity implements NavigationView.On
     private int correctAnswersCounter;
     private int totalPages;
     private Integer noOfQuestions;
+    DrawerAdapter drawerAdapter;
 
     //------For testing------
     final List<Card> questions = new ArrayList<>();
@@ -85,11 +84,10 @@ public class ExamActivity extends AppCompatActivity implements NavigationView.On
     private void initView() {
         ButterKnife.bind(this);
         setSupportActionBar(toolbar);
-        navigationView.setNavigationItemSelectedListener(this);
-        ActionBarDrawerToggle drawerToggle = new ActionBarDrawerToggle(
-                this, drawerLayout, toolbar, R.string.nav_open_drawer, R.string.nav_close_drawer);
-        drawerLayout.addDrawerListener(drawerToggle);
-        drawerToggle.syncState();
+        Context context = getApplicationContext();      //only For testing
+        drawerAdapter = new DrawerAdapter(navigationView, drawerLayout, toolbar, context);
+        drawerAdapter.attachDrawer();
+
         deckName.setText(deck.deckName);
         questionNo.setText(getString(R.string.question_no, pageCounter/2 + 1));
         correctAnswers.setText(getString(
@@ -178,34 +176,6 @@ public class ExamActivity extends AppCompatActivity implements NavigationView.On
             cards = crds;
         }
     }
-    @Override
-    public boolean onNavigationItemSelected(MenuItem item) {
-        item.setChecked(true);
-
-        int id = item.getItemId();
-        drawerLayout.closeDrawer(GravityCompat.START);
-
-        switch (id) {
-            case R.id.my_account:
-                break;
-            case R.id.my_decks:
-                break;
-            case R.id.create_flashcard:
-                break;
-            case R.id.show_deck:
-                break;
-            case R.id.statistics:
-                break;
-            case R.id.logout:
-                break;
-        }
-
-        // Only for testing
-        Toast.makeText(this, "Selected: " + item.getTitle(), Toast.LENGTH_SHORT).show();
-
-        return true;
-    }
-
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
